requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.7
django-appconf==1.0.4
django-imagekit==4.0.2
djangorestframework==3.11.0
gunicorn==20.0.4
install==1.3.3
pilkit==2.0
Pillow==7.2.0
pkg-resources==0.0.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
